fix(server): move phone-home server-id from package.json  to a seperated file and added the file to .gitignore #2486
