+++ packages/amplication-server/.gitignore
@@ -10,3 +10,4 @@ prisma/migrations/dev/
 artifacts/
 .env.local
 .eslintcache
+.server-id
